Begin 1.17 support
(from jglr)

+++ common/src/test/java/mca/MCASaving.java
@@ -1,47 +1,73 @@
 package mca;
 
+import org.jglrxavpok.hephaistos.data.DataSource;
+import org.jglrxavpok.hephaistos.data.GrowableSource;
 import org.jglrxavpok.hephaistos.mca.AnvilException;
 import org.jglrxavpok.hephaistos.mca.BlockState;
 import org.jglrxavpok.hephaistos.mca.ChunkColumn;
 import org.jglrxavpok.hephaistos.mca.RegionFile;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.function.Supplier;
 
-import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+@RunWith(Parameterized.class)
 public class MCASaving {
 
-    private RandomAccessFile file;
+    private DataSource dataSource;
+    private Supplier<DataSource> sourceSupplier;
+
+    @Parameterized.Parameters(name = "{index}")
+    public static Iterable<Object[]> data() {
+        Supplier<DataSource> growable = GrowableSource::new;
+        Supplier<DataSource> raf = () -> {
+            File target = new File("tmp_save_r.0.0.mca");
+            try {
+                target.createNewFile();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            try {
+                return new RandomAccessFileSource(new RandomAccessFile(target, "rw"));
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            }
+            return null;
+        };
+        return Arrays.asList(new Object[][]{
+                        {
+                                raf
+                        },
+                        {
+                                growable
+                        }
+                }
+        );
+    }
+
+    public MCASaving(Supplier<DataSource> sourceSupplier) {
+        this.sourceSupplier = sourceSupplier;
+    }
 
     @BeforeEach
     public void init() throws IOException {
-        File target = new File("tmp_save_r.0.0.mca");
-        target.createNewFile();
-        file = new RandomAccessFile(target, "rw");
-    }
-
-    @Test
-    public void loadAndSave() throws IOException, AnvilException {
-        file.close(); // don't need tmp file right now
-
-        Files.copy(Paths.get("src/test/resources/r.0.0.mca"), Paths.get("tmp_save_r.0.0.mca"), REPLACE_EXISTING);
-        file = new RandomAccessFile(new File("tmp_save_r.0.0.mca"), "rw");
-        RegionFile r = new RegionFile(file, 0, 0);
-        ChunkColumn chunk0_0 = r.getChunk(0,0);
-        assertNotNull(chunk0_0);
-        r.writeColumn(chunk0_0);
+        dataSource = sourceSupplier.get();
     }
 
     @Test
     public void verySimpleCreation() throws AnvilException, IOException {
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn newColumn = region.getOrCreateChunk(0, 0);
         assertEquals(ChunkColumn.GenerationStatus.Empty, newColumn.getGenerationStatus());
 
@@ -52,20 +78,20 @@ public class MCASaving {
     public void reuseFreeSector() throws AnvilException, IOException {
         final int sectionTotal = 10; // how many sections we want to use
         // purposely use too many to force having 8 free sectors
-        file.setLength(4096 * (sectionTotal+2));
-        file.seek(0);
+        dataSource.setLength(4096 * (sectionTotal+2));
+        dataSource.seek(0);
         // ensure file is full of zeroes
-        for (long i = 0; i < file.length(); i++) {
-            file.write(0);
+        for (long i = 0; i < dataSource.length(); i++) {
+            dataSource.write(0);
         }
-        file.seek(0);
+        dataSource.seek(0);
 
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn column = new ChunkColumn(0,0);
         region.writeColumn(column);
 
-        file.seek(0);
-        int location = file.readInt();
+        dataSource.seek(0);
+        int location = dataSource.readInt();
         int start = (location >> 8) & 0xFFFFFF;
         int length = location & 0xFF;
         assertEquals(2, start); // first free sector
@@ -74,8 +100,8 @@ public class MCASaving {
         region.writeColumn(column);
 
         // after a rewrite, should be moved to new sector
-        file.seek(0);
-        location = file.readInt();
+        dataSource.seek(0);
+        location = dataSource.readInt();
         start = (location >> 8) & 0xFFFFFF;
         length = location & 0xFF;
         assertEquals(3, start); // 2nd free sector
@@ -84,8 +110,8 @@ public class MCASaving {
         region.writeColumn(column);
 
         // after another rewrite, should be moved to the first sector, as it is now free
-        file.seek(0);
-        location = file.readInt();
+        dataSource.seek(0);
+        location = dataSource.readInt();
         start = (location >> 8) & 0xFFFFFF;
         length = location & 0xFF;
         assertEquals(2, start); // first free sector
@@ -96,20 +122,20 @@ public class MCASaving {
     public void useFreeSector() throws AnvilException, IOException {
         final int sectionTotal = 10; // how many sections we want to use
         // purposely use too many to force having 8 free sectors
-        file.setLength(4096 * (sectionTotal+2));
-        file.seek(0);
+        dataSource.setLength(4096 * (sectionTotal+2));
+        dataSource.seek(0);
         // ensure file is full of zeroes
-        for (long i = 0; i < file.length(); i++) {
-            file.write(0);
+        for (long i = 0; i < dataSource.length(); i++) {
+            dataSource.write(0);
         }
-        file.seek(0);
+        dataSource.seek(0);
 
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn column = new ChunkColumn(0,0);
         region.writeColumn(column);
 
-        file.seek(0);
-        int location = file.readInt();
+        dataSource.seek(0);
+        int location = dataSource.readInt();
         int start = (location >> 8) & 0xFFFFFF;
         int length = location & 0xFF;
         assertEquals(2, start); // first free sector
@@ -118,33 +144,35 @@ public class MCASaving {
 
     @Test
     public void allocateNewSector() throws AnvilException, IOException {
-        file.setLength(4096 * 2);
-        file.seek(0);
+        dataSource.setLength(4096 * 2);
+        dataSource.seek(0);
         // ensure file is full of zeroes
-        for (long i = 0; i < file.length(); i++) {
-            file.write(0);
+        for (long i = 0; i < dataSource.length(); i++) {
+            dataSource.write(0);
         }
-        assertEquals(4096* 2, file.length()); // no free sectors
-        file.seek(0);
+        assertEquals(4096* 2, dataSource.length()); // no free sectors
+        dataSource.seek(0);
 
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn column = new ChunkColumn(0,0);
         region.writeColumn(column);
 
-        assertEquals(4096*3, file.length()); // allocated one new sector
+        assertEquals(4096*3, dataSource.length()); // allocated one new sector
 
-        file.seek(0);
-        int location = file.readInt();
+        dataSource.seek(0);
+        int location = dataSource.readInt();
         int start = (location >> 8) & 0xFFFFFF;
         int length = location & 0xFF;
         assertEquals(2, start); // first free sector
         assertEquals(1, length); // chunk very light, should only use a single 4kib sector
     }
 
+    // TODO: 1.17 TESTS
+
     @Test
     public void creationFromScratchViaRegionFile() throws AnvilException, IOException {
         {
-            RegionFile region = new RegionFile(file, 0, 0);
+            RegionFile region = new RegionFile(dataSource, 0, 0);
             BlockState stone = new BlockState("minecraft:stone");
             for (int x = 0; x < 16; x++) {
                 for (int z = 0; z < 16; z++) {
@@ -158,7 +186,7 @@ public class MCASaving {
         }
 
         {
-            RegionFile region = new RegionFile(file, 0, 0);
+            RegionFile region = new RegionFile(dataSource, 0, 0);
             BlockState stone = new BlockState("minecraft:stone");
             for (int x = 0; x < 16; x++) {
                 for (int z = 0; z < 16; z++) {
@@ -174,40 +202,40 @@ public class MCASaving {
     @Test
     public void creationFromScratchViaChunks() throws AnvilException, IOException {
         {
-            RegionFile region = new RegionFile(file, 0, 0);
+            RegionFile region = new RegionFile(dataSource, 0, 0);
             ChunkColumn chunk0 = region.getOrCreateChunk(0, 0);
             ChunkColumn chunk1 = region.getOrCreateChunk(1, 0);
-            fillChunk(chunk0, BlockState.AIR);
-            fillChunk(chunk1, new BlockState("minecraft:stone"));
+            fillChunk(chunk0, BlockState.AIR, 0, 255);
+            fillChunk(chunk1, new BlockState("minecraft:stone"), 0, 255);
             region.writeColumn(chunk0);
             region.writeColumn(chunk1);
         }
 
         {
-            RegionFile reloaded = new RegionFile(file, 0, 0);
+            RegionFile reloaded = new RegionFile(dataSource, 0, 0);
             ChunkColumn reloadedChunk0 = reloaded.getChunk(0, 0);
             ChunkColumn reloadedChunk1 = reloaded.getChunk(1, 0);
             assertNotNull(reloadedChunk0);
             assertNotNull(reloadedChunk1);
-            verifyChunk(reloadedChunk0, BlockState.AIR);
-            verifyChunk(reloadedChunk1, new BlockState("minecraft:stone"));
+            verifyChunk(reloadedChunk0, BlockState.AIR, 0, 255);
+            verifyChunk(reloadedChunk1, new BlockState("minecraft:stone"), 0, 255);
         }
     }
 
-    private void verifyChunk(ChunkColumn chunkColumn, BlockState state) {
+    private void verifyChunk(ChunkColumn chunkColumn, BlockState state, int minY, int maxY) {
         for (int x = 0; x < 16; x++) {
             for (int z = 0; z < 16; z++) {
-                for (int y = 0; y < 256; y++) {
+                for (int y = minY; y <= maxY; y++) {
                     assertEquals(state, chunkColumn.getBlockState(x, y, z));
                 }
             }
         }
     }
 
-    private void fillChunk(ChunkColumn chunkColumn, BlockState state) {
+    private void fillChunk(ChunkColumn chunkColumn, BlockState state, int minY, int maxY) {
         for (int x = 0; x < 16; x++) {
             for (int z = 0; z < 16; z++) {
-                for (int y = 0; y < 256; y++) {
+                for (int y = minY; y <= maxY; y++) {
                     chunkColumn.setBlockState(x, y, z, state);
                 }
             }
@@ -216,7 +244,7 @@ public class MCASaving {
 
     @Test
     public void setBiomes() throws AnvilException, IOException {
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         region.getOrCreateChunk(0, 0); // force create chunk
         assertEquals(ChunkColumn.UnknownBiome, region.getBiome(0,0,0));
         region.setBiome(0,0,0, 1);
@@ -225,31 +253,88 @@ public class MCASaving {
 
     @Test
     public void setBlockLight() throws AnvilException, IOException {
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn c = region.getOrCreateChunk(0, 0);
-        c.getSections()[0].setBlockLight(0,0,0, (byte) 15);
-        c.getSections()[0].setBlockLight(0,5,0, (byte) 12);
-        assertEquals((byte) 15, c.getSections()[0].getBlockLight(0,0,0));
-        assertEquals((byte) 12, c.getSections()[0].getBlockLight(0,5,0));
+        c.getSection((byte)0).setBlockLight(0,0,0, (byte) 15);
+        c.getSection((byte)0).setBlockLight(0,5,0, (byte) 12);
+        assertEquals((byte) 15, c.getSection((byte)0).getBlockLight(0,0,0));
+        assertEquals((byte) 12, c.getSection((byte)0).getBlockLight(0,5,0));
     }
 
     @Test
     public void setSkyLight() throws AnvilException, IOException {
-        RegionFile region = new RegionFile(file, 0, 0);
+        RegionFile region = new RegionFile(dataSource, 0, 0);
         ChunkColumn c = region.getOrCreateChunk(0, 0);
-        c.getSections()[0].setSkyLight(0,0,0, (byte) 15);
-        c.getSections()[0].setSkyLight(0,5,0, (byte) 12);
-        assertEquals((byte) 15, c.getSections()[0].getSkyLight(0,0,0));
-        assertEquals((byte) 12, c.getSections()[0].getSkyLight(0,5,0));
+        c.getSection((byte)0).setSkyLight(0,0,0, (byte) 15);
+        c.getSection((byte)0).setSkyLight(0,5,0, (byte) 12);
+        assertEquals((byte) 15, c.getSection((byte)0).getSkyLight(0,0,0));
+        assertEquals((byte) 12, c.getSection((byte)0).getSkyLight(0,5,0));
     }
+
+    // Start of 1.17 support ============================================
+
+    @Test
+    public void creationFromScratchViaChunks_1_17() throws AnvilException, IOException {
+        int[] randomValues = {
+                -2598, -267, -428, -392, 899,
+                32, 64, 128, 0, 256, 31, 63, 127, 255,
+                342, 341
+        };
+
+        long progress = 0;
+        final long total = randomValues.length * randomValues.length;
+
+        for (int minY : randomValues) {
+            for (int maxY : randomValues) {
+                boolean expectAnvilException = minY > maxY;
+                try {
+                    {
+                        RegionFile region = new RegionFile(dataSource, 0, 0, minY, maxY);
+                        ChunkColumn chunk0 = region.getOrCreateChunk(0, 0);
+                        ChunkColumn chunk1 = region.getOrCreateChunk(1, 0);
+                        fillChunk(chunk0, BlockState.AIR, minY, maxY);
+                        fillChunk(chunk1, new BlockState("minecraft:stone"), minY, maxY);
+                        region.writeColumn(chunk0);
+                        region.writeColumn(chunk1);
+                    }
+
+                    {
+                        RegionFile reloaded = new RegionFile(dataSource, 0, 0, minY, maxY);
+                        ChunkColumn reloadedChunk0 = reloaded.getChunk(0, 0);
+                        ChunkColumn reloadedChunk1 = reloaded.getChunk(1, 0);
+                        assertNotNull(reloadedChunk0);
+                        assertNotNull(reloadedChunk1);
+                        verifyChunk(reloadedChunk0, BlockState.AIR, minY, maxY);
+                        verifyChunk(reloadedChunk1, new BlockState("minecraft:stone"), minY, maxY);
+                    }
+                } catch (AnvilException e) {
+                    if(!expectAnvilException || !e.getMessage().equals("minY must be <= maxY")) {
+                        throw e;
+                    }
+                }
+
+                if(progress % 10 == 0) {
+                    System.out.println("In progress... " + progress + " / " + total + "");
+                }
+                progress++;
+            }
+        }
+
+        System.out.println("Done! (" + progress + " / " + total + ")");
+    }
+
+
+    // End of 1.17 support ============================================
 
     @AfterEach
     public void clean() throws IOException {
         try {
-            file.close();
+            dataSource.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
-        Files.delete(Paths.get("tmp_save_r.0.0.mca"));
+
+        if(dataSource instanceof RandomAccessFileSource)
+            Files.delete(Paths.get("tmp_save_r.0.0.mca"));
     }
 }
